requirements.txt
aiosqlite==0.19.0
alembic==1.11.1
anyio==3.7.1
apprise==1.4.5
asgi-lifespan==2.1.0
asyncpg==0.28.0
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.1.0
coolname==2.2.0
croniter==1.4.1
cryptography==41.0.2
cycler==0.11.0
databricks-cli==0.17.7
dateparser==1.1.8
db-dtypes==1.1.1
distlib==0.3.7
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.12.2
Flask==2.3.2
fonttools==4.41.1
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.11.4
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
griffe==0.32.3
grpcio==1.56.2
grpcio-status==1.56.2
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
ibm-cos-sdk==2.11.0
ibm-cos-sdk-core==2.11.0
ibm-cos-sdk-s3transfer==2.11.0
ibm-watson-machine-learning==1.0.311
identify==2.5.26
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
kubernetes==27.2.0
lomond==0.3.3
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mlflow==2.5.0
nodeenv==1.8.0
numpy==1.25.1
oauthlib==3.2.2
orjson==3.9.2
packaging==23.1
pandas==1.5.3
pandas-gbq==0.19.2
pathspec==0.11.1
pendulum==2.1.2
Pillow==10.0.0
pipenv==2023.7.11
platformdirs==3.9.1
pluggy==1.2.0
pre-commit==3.3.3
prefect==2.11.0
proto-plus==1.22.3
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
pydata-google-auth==1.8.1
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
pytest==7.4.0
pytest-dependency==0.5.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
querystring-parser==1.2.4
readchar==4.0.5
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
starlette==0.27.0
tabulate==0.9.0
text-unidecode==1.3
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
typer==0.9.0
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==1.26.16
uvicorn==0.23.1
virtualenv==20.24.2
virtualenv-clone==0.5.7
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
zipp==3.16.2
